requirements/base.txt
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.23
celery==3.1.23
defusedxml==0.4.1
Django==1.9.6
django-braces==1.8.1
django-debug-toolbar==1.4
django-push-notifications==1.4.1
django-rest-swagger==0.3.7
djangorestframework==3.3.3
djangorestframework-jwt==1.8.0
drfdocs==0.0.11
ecdsa==0.13
enum34==1.1.4
Fabric3==1.11.1.post1
greenlet==0.4.7
hashids==1.1.0
idna==2.1
ipaddress==1.0.16
kombu==3.0.35
oauthlib==1.1.1
paramiko==1.17.0
Pillow==5.2.0
psycopg2-binary==2.7.5
pyasn1==0.1.9
pycparser==2.14
pycrypto==2.6.1
PyJWT==1.4.0
python-openid==2.2.5
python-social-auth==0.2.19
python3-openid==3.0.10
pytz==2016.4
PyYAML==3.11
qrcode==5.1
requests==2.10.0
requests-oauthlib==0.6.1
six==1.10.0
sqlparse==0.1.16
